fix bug in groove midi download

+++ mirdata/datasets/groove_midi.py
@@ -488,9 +488,25 @@ class Dataset(core.Dataset):
 
         # files get downloaded to a folder called groove - move everything up a level
         groove_dir = os.path.join(self.data_home, "groove")
+        if not os.path.exists(groove_dir):
+            logging.info(
+                "Groove MIDI data not downloaded, because it probably already exists on your computer. "
+                + "Run .validate() to check, or rerun with force_overwrite=True to delete any "
+                + "existing files and download from scratch"
+            )
+            return
+
         groove_files = glob.glob(os.path.join(groove_dir, "*"))
 
         for fpath in groove_files:
+            target_path = os.path.join(self.data_home, os.path.basename(fpath))
+            if os.path.exists(target_path):
+                logging.info(
+                    "{} already exists. Run with force_overwrite=True to download from scratch".format(
+                        target_path
+                    )
+                )
+                continue
             shutil.move(fpath, self.data_home)
 
         if os.path.exists(groove_dir):
